refactor(layer-selector): refactor stories with map into a separate module
